refactor: shifted payments pkg entirely to api
