--- internal/payments/utils.go
@@ -1,14 +1,0 @@
-package payments
-
-import (
-	"encoding/json"
-)
-
-func MapToStruct(source any, target interface{}) error {
-	entity, err := json.Marshal(source)
-	if err != nil {
-		return err
-	}
-
-	return json.Unmarshal(entity, &target)
-}
