+++ main.go
@@ -9,12 +9,12 @@ import (
 	"github.com/envsecrets/envsecrets/api/environments"
 	"github.com/envsecrets/envsecrets/api/events"
 	"github.com/envsecrets/envsecrets/api/integrations"
+	"github.com/envsecrets/envsecrets/api/payments"
 	"github.com/envsecrets/envsecrets/internal/actions"
 	"github.com/envsecrets/envsecrets/internal/auth"
 	"github.com/envsecrets/envsecrets/internal/invites"
 	"github.com/envsecrets/envsecrets/internal/keys"
 	"github.com/envsecrets/envsecrets/internal/middlewares"
-	"github.com/envsecrets/envsecrets/internal/payments"
 	"github.com/envsecrets/envsecrets/internal/secrets"
 	"github.com/envsecrets/envsecrets/internal/tokens"
 	"github.com/envsecrets/envsecrets/internal/triggers"
